[MME][Memory] Enable UBSAN for debug builds and resolve unaligned memory access issues (#5262)

+++ lte/gateway/c/oai/tasks/sgw/sgw_handlers.c
@@ -1896,10 +1896,11 @@ static void _generate_dl_flow(
   if ((TRAFFIC_FLOW_TEMPLATE_IPV4_REMOTE_ADDR_FLAG & packet_filter->flags) ==
       TRAFFIC_FLOW_TEMPLATE_IPV4_REMOTE_ADDR_FLAG) {
     struct in_addr remoteaddr = {.s_addr = 0};
-    remoteaddr.s_addr         = (packet_filter->ipv4remoteaddr[0].addr << 24) +
-                        (packet_filter->ipv4remoteaddr[1].addr << 16) +
-                        (packet_filter->ipv4remoteaddr[2].addr << 8) +
-                        packet_filter->ipv4remoteaddr[3].addr;
+    remoteaddr.s_addr =
+        (((uint32_t) packet_filter->ipv4remoteaddr[0].addr) << 24) +
+        (((uint32_t) packet_filter->ipv4remoteaddr[1].addr) << 16) +
+        (((uint32_t) packet_filter->ipv4remoteaddr[2].addr) << 8) +
+        (((uint32_t) packet_filter->ipv4remoteaddr[3].addr));
     dlflow->src_ip.s_addr = ntohl(remoteaddr.s_addr);
     dlflow->set_params |= SRC_IPV4;
   }
